R6 base class for authentication schemes

All schemes inherit from this base class and adapts it for the particular scheme it implements. Additional schemes can be implemented as subclasses of this and will work transparently with fireproof.

is_auth(x)

Arguments

x

An object

Active bindings

name

The name of the instance

open_api

An OpenID compliant security scheme description

Methods

Public methods

Auth$new()

Auth$check_request()

Auth$reject_response()

Auth$forbid_user()

Auth$clone()

Method new()

Constructor for the class

Usage

Auth$new(name = NULL)

Arguments

name

The name of the scheme instance

Method check_request()

A function that validates an incoming request, returning TRUE if it is valid and FALSE if not. The base class simply returns TRUE for all requests

Usage

Auth$check_request(request, response, keys, ...)

Arguments

request

The request to validate as a Request object

response

The corresponding response to the request as a Response object

keys

A named list of path parameters from the path matching

...

Ignored

Method reject_response()

Action to perform on the response in case the request fails to get validated by any instance in the flow. All failing instances will have this method called one by one so be mindful if you are overwriting information set by another instance

Usage

Auth$reject_response(response, scope)

Arguments

response

The response object

scope

The scope of the endpoint

Method forbid_user()

Action to perform on the response in case the request does not have the necessary permissions for the endpoint. All succeeding instances will have this method called one by one if permissions are insufficient so be mindful if you are overwriting information set by another instance

Usage

Auth$forbid_user(response, scope)

Arguments

response

The response object

scope

The scope of the endpoint

Method clone()

The objects of this class are cloneable with this method.

Usage

Auth$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.