R6 class for the Basic authentication scheme

This class encapsulates the logic of the Basic authentication scheme. See auth_basic() for more information.

Super class

fireproof::Auth -> AuthBasic

Methods

Public methods

AuthBasic$new()

AuthBasic$check_request()

AuthBasic$reject_response()

AuthBasic$clone()

Inherited methods

fireproof::Auth$forbid_user()

Method new()

Constructor for the class

Usage

AuthBasic$new(authenticator, realm = "private", name = NULL)

Arguments

authenticator

A function that takes a username, password, realm, request, and response and returns TRUE if the pair is valid, and FALSE otherwise. If the function sets a character vector of scopes for the user in the auth_scope field of the response data (response$set_data("auth_scope", ...)) then it will be tested against the scopes needed for the specific endpoint

realm

The realm this authentication corresponds to. Will be returned to the client on a failed authentication attempt to inform them of the credentials required, though most often these days it is kept from the user.

name

The name of the authentication

Method check_request()

A function that validates an incoming request, returning TRUE if it is valid and FALSE if not. It decodes the credentials in the Authorization header, splits it into username and password and then calls the authenticator function provided at construction.

Usage

AuthBasic$check_request(request, response, keys, ...)

Arguments

request

The request to validate as a Request object

response

The corresponding response to the request as a Response object

keys

A named list of path parameters from the path matching

...

Ignored

Method reject_response()

Upon rejection this scheme sets the response status to 401 and sets the WWW-Authenticate header to Basic realm="<realm>", charset=UTF-8

Usage

AuthBasic$reject_response(response, scope)

Arguments

response

The response object

scope

The scope of the endpoint

Method clone()

The objects of this class are cloneable with this method.

Usage

AuthBasic$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.