R6 class for the Bearer authentication scheme

This class encapsulates the logic of the Bearer authentication scheme. See auth_bearer() for more information.

Super class

fireproof::Auth -> AuthBearer

Active bindings

open_api

An OpenID compliant security scheme description

Methods

Public methods

AuthBearer$new()

AuthBearer$check_request()

AuthBearer$reject_response()

AuthBearer$clone()

Inherited methods

fireproof::Auth$forbid_user()

fireproof::Auth$register_handler()

Method new()

Constructor for the class

Usage

AuthBearer$new(
  authenticator,
  user_info = NULL,
  realm = "private",
  allow_body_token = TRUE,
  allow_query_token = FALSE,
  name = NULL
)

Arguments

authenticator

A function that will be called with the arguments token, realm, request, and response and returns TRUE if the token is valid, and FALSE otherwise. If the function returns a character vector it is considered to be authenticated and the return value will be understood as scopes the user is granted.

user_info

A function to extract user information from the username. It is called with two arguments: token and setter, the first being the token used for the successful authentication, the second being a function that must be called in the end with the relevant information (see user_info()).

realm

The realm this authentication corresponds to. Will be returned to the client on a failed authentication attempt to inform them of the credentials required, though most often these days it is kept from the user.

allow_body_token

Should it be allowed to pass the token in the request body as a query form type with the access_token name. Defaults to TRUE but you can turn it off to force the client to use the Authorization header.

allow_query_token

Should it be allowed to pass the token in the query string of the url with the access_token name. Default to FALSE due to severe security implications but can be turned on if you have very well-thought-out reasons to do so.

name

The name of the authentication

Method check_request()

A function that validates an incoming request, returning TRUE if it is valid and FALSE if not. It fetches the token from the request according to the allow_body_token and allow_query_token settings and validates it according to the provided authenticator. If the token is present multiple times it will fail with 400 is this is not allowed. It will store the token in the auth_token field of the response data.

Usage

AuthBearer$check_request(request, response, keys, ..., .session)

Arguments

request

The request to validate as a Request object

response

The corresponding response to the request as a Response object

keys

A named list of path parameters from the path matching

...

Ignored

.session

The session storage for the current session

Method reject_response()

Upon rejection this scheme sets the response status to 401 and sets the WWW-Authenticate header to Bearer realm="<realm>". If any scope is provided by the endpoint it will be appended as , scope="<scope>" and if the token is present but invalid, it will append , error="invalid_token"

Usage

AuthBearer$reject_response(response, scope, ..., .session)

Arguments

response

The response object

scope

The scope of the endpoint

...

Ignored

.session

The session storage for the current session

Method clone()

The objects of this class are cloneable with this method.

Usage

AuthBearer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Create an authenticator of dubious quality
bearer <- AuthBearer$new(
  authenticator = function(token) {
    token == "abcd1234"
  },
  user_info = function(user, setter) {
    setter(
      name_given = "Thomas",
      name_middle = "Lin",
      name_family = "Pedersen"
    )
  }
)