R6 class for the Key authentication scheme

This class encapsulates the logic of the key based authentication scheme. See auth_key() for more information

Super class

fireproof::Auth -> AuthKey

Active bindings

location

The location of the secret in the request, either "cookie" or "header"

open_api

An OpenID compliant security scheme description

Methods

Public methods

AuthKey$new()

AuthKey$check_request()

AuthKey$reject_response()

AuthKey$clone()

Inherited methods

fireproof::Auth$forbid_user()

fireproof::Auth$register_handler()

Method new()

Constructor for the class

Usage

AuthKey$new(key, secret, user_info = NULL, cookie = TRUE, name = NULL)

Arguments

key

The name of the header or cookie to store the secret under

secret

The secret to check for. Either a single string with the secret or a function that will be called with the arguments key, request, and response and returns TRUE if its a valid secret (useful if you have multiple or rotating secrets). If the function returns a character vector it is considered to be authenticated and the return value will be understood as scopes the user is granted. Make sure never to store secrets in plain text and avoid checking them into version control.

user_info

A function to extract user information from the username. It is called with two arguments: token and setter, the first being the key used for the successful authentication, the second being a function that must be called in the end with the relevant information (see user_info()).

cookie

Boolean. Should the secret be transmitted as a cookie. If FALSE it is expected to be transmitted as a header.

name

The name of the scheme instance

Method check_request()

A function that validates an incoming request, returning TRUE if it is valid and FALSE if not. It extracts the secret from either the cookie or header based on the provided key and test it against the provided secret.

Usage

AuthKey$check_request(request, response, keys, ..., .session)

Arguments

request

The request to validate as a Request object

response

The corresponding response to the request as a Response object

keys

A named list of path parameters from the path matching

...

Ignored

.session

The session storage for the current session

server

The fiery server handling the request

arg_list

A list of additional arguments extracted be the before_request handlers (will be used to access the session data store)

Method reject_response()

Upon rejection this scheme sets the response status to 400 if it has not already been set by others. In contrast to the other schemes that are proper HTTP schemes, this one doesn't set a WWW-Authenticate header.

Usage

AuthKey$reject_response(response, scope, ..., .session)

Arguments

response

The response object

scope

The scope of the endpoint

...

Ignored

.session

The session storage for the current session

Method clone()

The objects of this class are cloneable with this method.

Usage

AuthKey$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Create an authenticator of dubious quality
key <- auth_key(
  key = "my-key-location",
  secret = "SHHH!!DONT_TELL_ANYONE"
)