R6 class for the OAuth 2.0 authentication scheme

This class encapsulates the logic of the oauth 2.0 based authentication scheme. See auth_oauth2() for more information

Super class

fireproof::Auth -> AuthOAuth2

Active bindings

location

The location of the secret in the request, either "cookie" or "header"

open_api

An OpenID compliant security scheme description

Methods

Public methods

AuthOAuth2$new()

AuthOAuth2$check_request()

AuthOAuth2$reject_response()

AuthOAuth2$register_handler()

AuthOAuth2$clone()

Inherited methods

fireproof::Auth$forbid_user()

Method new()

Constructor for the class

Usage

AuthOAuth2$new(
  token_url,
  redirect_url,
  client_id,
  client_secret,
  auth_url = NULL,
  grant_type = c("authorization_code", "password"),
  scopes = NULL,
  validate = function(info) TRUE,
  redirect_path = sub("^.*?(?=(?<!/)/(?!/))", "", redirect_url),
  on_auth = replay_request,
  user_info = NULL,
  service_params = list(),
  name = NULL
)

Arguments

token_url

The URL to the authorization servers token endpoint

redirect_url

The URL the authorization server should redirect to following a successful authorization. Must be equivalent to one provided when registering your application

client_id

The ID issued by the authorization server when registering your application

client_secret

The secret issued by the authorization server when registering your application. Do NOT store this in plain text

auth_url

The URL to redirect the user to when requesting authorization (only needed for grant_type = "authorization_code")

grant_type

The type of authorization scheme to use, either "authorization_code" or "password"

scopes

Optional character vector of scopes to request the user to grant you during authorization

validate

Function to validate the user once logged in. It will be called with a single argument info, which gets the information of the user as provided by the user_info function in the. By default it returns TRUE on everything meaning that anyone who can log in with the provider will be accepted, but you can provide a different function to e.g. restrict access to certain user names etc.

redirect_path

The path that should capture redirects after successful authorization. By default this is derived from redirect_url by removing the domain part of the url, but if for some reason this doesn't yields the correct result for your server setup you can overwrite it here.

on_auth

A function which will handle the result of a successful authorization. It will be called with four arguments: request, response, session_state, and server. The first contains the current request being responded to, the second is the response being send back, the third is a list recording the state of the original request which initiated the authorization (containing method, url, headers, and body fields with information from the original request). By default it will use replay_request to internally replay the original request and send back the response.

user_info

A function to extract user information from the authorization provider. It will be called with two arguments: token_info and setter, the first being the token information returned from the provider as a list (notably with a token field for the actual token), the second being a function that must be called in the end with the relevant information. The setter function takes the following arguments: provider (the name of the oauth2 provider), id (the identifier of the user), display_name (the name the user has chosen as public name), name_given (the users real given name), name_middle (the users middle name), name_family (the users family name), emails (a vector of emails, potentially named with type, e.g. "work", "home" etc), photos (a vector of urls for profile photos), and ... with additional named fields to add

service_params

A named list of additional query params to add to the url when constructing the authorization url in the "authorization_code" grant type

name

The name of the scheme instance. This will also be the name under which token info and user info is saved in the session store

Method check_request()

A function that validates an incoming request, returning TRUE if it is valid and FALSE if not. It extracts the secret from either the cookie or header based on the provided key and test it against the provided secret.

Usage

AuthOAuth2$check_request(request, response, keys, ..., .session)

Arguments

request

The request to validate as a Request object

response

The corresponding response to the request as a Response object

keys

A named list of path parameters from the path matching

...

Ignored

.session

The session storage for the current session

server

The fiery server handling the request

arg_list

A list of additional arguments extracted be the before_request handlers (will be used to access the session data store)

Method reject_response()

Upon rejection this scheme sets the response status to 400 if it has not already been set by others. In contrast to the other schemes that are proper HTTP schemes, this one doesn't set a WWW-Authenticate header.

Usage

AuthOAuth2$reject_response(response, scope, ..., .session)

Arguments

response

The response object

scope

The scope of the endpoint

...

Ignored

.session

The session storage for the current session

Method register_handler()

Hook for registering endpoint handlers needed for this authentication method

Usage

AuthOAuth2$register_handler(add_handler)

Arguments

add_handler

The add_handler method from Fireproof to be called for adding additional handlers

Method clone()

The objects of this class are cloneable with this method.

Usage

AuthOAuth2$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Example using GitHub endpoints (use `auth_github()` in real code)
github <- AuthOAuth2$new(
  token_url = "https://github.com/login/oauth/access_token",
  redirect_url = "https://example.com/auth",
  client_id = "MY_APP_ID",
  client_secret = "SUCHASECRET",
  auth_url = "https://github.com/login/oauth/authorize",
  grant_type = "authorization_code"
)
#> Warning: TRE pattern compilation error 'Invalid regexp'
#> Error in sub("^.*?(?=(?<!/)/(?!/))", "", redirect_url): invalid regular expression '^.*?(?=(?<!/)/(?!/))', reason 'Invalid regexp'