Authenticate using the mock OAuth servers provided by Beeceptor

These two functions sets up mock OAuth 2.0 authentication based on tools provided by Beeceptor. These should obviously not be used for production because they allow anyone to be authenticated, but they can be used while testing your authentication setup.

auth_beeceptor_github(
  redirect_url,
  client_id = "MOCK_CLIENT",
  ...,
  name = "beeceptor_github"
)

auth_beeceptor_google(
  redirect_url,
  client_id = "MOCK_CLIENT",
  ...,
  name = "beeceptor_google"
)

Arguments

redirect_url

The URL the authorization server should redirect to following a successful authorization. Must be equivalent to one provided when registering your application

client_id

The ID issued by the authorization server when registering your application

...

Arguments passed on to auth_oauth2

token_url

The URL to the authorization servers token endpoint

client_secret

The secret issued by the authorization server when registering your application. Do NOT store this in plain text

auth_url

The URL to redirect the user to when requesting authorization (only needed for grant_type = "authorization_code")

grant_type

The type of authorization scheme to use, either "authorization_code" or "password"

scopes

Optional character vector of scopes to request the user to grant you during authorization

validate

Function to validate the user once logged in. It must have a single argument info, which gets the information of the user as provided by the user_info function in the. By default it returns TRUE on everything meaning that anyone who can log in with the provider will be accepted, but you can provide a different function to e.g. restrict access to certain user names etc.

redirect_path

The path that should capture redirects after successful authorization. By default this is derived from redirect_url by removing the domain part of the url, but if for some reason this doesn't yields the correct result for your server setup you can overwrite it here.

on_auth

A function which will handle the result of a successful authorization. It must have four arguments: request, response, session_state, and server. The first contains the current request being responded to, the second is the response being send back, the third is a list recording the state of the original request which initiated the authorization (containing method, url, headers, and body fields with information from the original request). By default it will use replay_request to internally replay the original request and send back the response.

user_info

A function to extract user information from the authorization provider. It takes two arguments: token_info and setter, the first being the token information returned from the provider as a list (notably with a token field for the actual token), the second being a function that must be called in the end with the relevant information. The setter function takes the following arguments: provider (the name of the oauth2 provider), id (the identifier of the user), display_name (the name the user has chosen as public name), name_given (the users real given name), name_middle (the users middle name), name_family (the users family name), emails (a vector of emails, potentially named with type, e.g. "work", "home" etc), photos (a vector of urls for profile photos), and ... with additional named fields to add

service_params

A named list of additional query params to add to the url when constructing the authorization url in the "authorization_code" grant type

name

The name of the scheme instance. This will also be the name under which token info and user info is saved in the session store

Value

An AuthOAuth2 object