Authenticate with Google using OpenID Connect

This authentication requests you to log in with google and authenticates you through their service. Your server must be registered and have a valid client ID and client secret for this to work. Read more about registering an application at https://developers.google.com/identity/protocols/oauth2. If you want to limit access to only select users you should make sure to provide a validation function that checks the userinfo against a whitelist.

auth_google(
  redirect_url,
  client_id,
  client_secret,
  scopes = "profile",
  service_params = list(access_type = "offline", include_granted_scopes = "true"),
  ...,
  name = "google"
)

Arguments

redirect_url

The URL the authorization server should redirect to following a successful authorization. Must be equivalent to one provided when registering your application

client_id

The ID issued by the authorization server when registering your application

client_secret

The secret issued by the authorization server when registering your application. Do NOT store this in plain text

scopes

Optional character vector of scopes to request the user to grant you during authorization

service_params

A named list of additional query params to add to the url when constructing the authorization url in the "authorization_code" grant type

...

Arguments passed on to auth_oidc

request_user_info

Logical. Should the userinfo endpoint be followed to add information about the user not present in the JWT token. Setting this to TRUE will add an additional API call to your authentication flow but potentially provide richer information about the user.

validate

Function to validate the user once logged in. It must have a single argument info, which gets the information of the user as provided by the user_info function in the. By default it returns TRUE on everything meaning that anyone who can log in with the provider will be accepted, but you can provide a different function to e.g. restrict access to certain user names etc.

redirect_path

The path that should capture redirects after successful authorization. By default this is derived from redirect_url by removing the domain part of the url, but if for some reason this doesn't yields the correct result for your server setup you can overwrite it here.

on_auth

A function which will handle the result of a successful authorization. It must have four arguments: request, response, session_state, and server. The first contains the current request being responded to, the second is the response being send back, the third is a list recording the state of the original request which initiated the authorization (containing method, url, headers, and body fields with information from the original request). By default it will use replay_request to internally replay the original request and send back the response.

name

The name of the scheme instance. This will also be the name under which token info and user info is saved in the session store

Value

An AuthOIDC object