A plugin that handles authentication

This plugin orchestrates all authentication for your fiery app. It is a special Route that manages all the different authentication scheme instances you have defined as well as testing all the endpoints that have authentication requirements.

Details

An authentication scheme instance is an object deriving from the Auth class which is usually created with one of the auth_*() constructors. You can provide it with a name as you register it and can thus have multiple instances of the same scheme (e.g. two auth_basic() with different user lists)

An authentication handler is a handler that consists of a method, path, and flow. The flow is a logical expression of the various instances the request must pass to get access to that endpoint. For example, if you have two instances named auth1 and auth2, a flow could be auth1 || auth2 to allow a request if it passes either of the instances. Given an additional instance, auth3, it could also be something like auth1 || (auth2 && auth3). The flow is given as a bare expression, not as a string. In addition to the three required arguments you can also supply a character vector of scopes that are required to be had to access the endpoint. If your scheme has scope support then the request will be tested against these to see if the (otherwise valid) user has permission to the resource.

Super class

routr::Route -> Fireproof

Active bindings

name

The name of the plugin: "fireproof"

require

Required plugins for Fireproof

auths

The name of all the authentication schemes currently added to the plugin

Methods

Public methods

Fireproof$add_auth_handler()

Fireproof$add_auth()

Fireproof$add_handler()

Fireproof$flow_to_openapi()

Fireproof$on_attach()

Fireproof$clone()

Inherited methods

routr::Route$dispatch()

routr::Route$get_handler()

routr::Route$initialize()

routr::Route$merge_route()

routr::Route$print()

routr::Route$remap_handlers()

routr::Route$remove_handler()

Method add_auth_handler()

Add a new authentication handler. It invisibly returns the parsed flow so it can be used for generating OpenAPI specs with.

Usage

Fireproof$add_auth_handler(method, path, flow, scope = NULL)

Arguments

method

The http method to match the handler to

path

The URL path to match to

flow

The authentication flow the request must pass to be valid. See Details

scope

An optional character vector of scopes that the request must have permission for to access the resource

Method add_auth()

Add an authentication scheme instance to the plugin

Usage

Fireproof$add_auth(auth, name = NULL)

Arguments

auth

Either an Auth object defining the scheme instance (preferred) or a function taking the standard route handler arguments (request, response, keys, and ...) and returns TRUE if the request is valid and FALSE if not.

name

The name of the scheme instance to be used when defining flow for endpoint authentication.

Method add_handler()

Defunct overwrite of the add_handler() method to prevent this route to be used for anything other than authentication. Will throw an error if called.

Usage

Fireproof$add_handler(method, path, handler, reject_missing_methods = FALSE)

Arguments

method

ignored

path

ignored

handler

ignored

reject_missing_methods

ignored

Method flow_to_openapi()

Turns a parsed flow (as returned by add_auth_handler()) into an OpenAPI Security Requirement compliant list. Not all flows can be represented by the OpenAPI spec and the method will return NULL with a warning if so. Scope is added to all schemes, even if not applicable, so the final OpenAPI doc should be run through prune_openapi() before serving it.

Usage

Fireproof$flow_to_openapi(flow, scope)

Arguments

flow

A parsed flow as returned by add_auth_handler()

scope

A character vector of scopes required for this particular flow

Method on_attach()

Method for use by fiery when attached as a plugin. Should not be called directly. This method looks for a header route stack in the app and if it doesn't exist it creates one. It then attaches the plugin as the first route to it.

Usage

Fireproof$on_attach(app, ...)

Arguments

app

The Fire object to attach the router to

...

Ignored

Method clone()

The objects of this class are cloneable with this method.

Usage

Fireproof$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Create a fireproof plugin
fp <- Fireproof$new()

# Create some authentication schemes and add them
basic <- auth_basic(
  authenticator = function(user, password) {
    user == "thomas" && password == "pedersen"
  },
  user_info = function(user, setter) {
    setter(
      name_given = "Thomas",
      name_middle = "Lin",
      name_family = "Pedersen"
    )
  }
)
fp$add_auth(basic, "basic_auth")

key <- auth_key(
  key = "my-key-location",
  secret = "SHHH!!DONT_TELL_ANYONE"
)
fp$add_auth(key, "key_auth")

google <- auth_google(
  redirect_url = "https://example.com/auth",
  client_id = "MY_APP_ID",
  client_secret = "SUCHASECRET",
)
#> Warning: PCRE pattern compilation error
#> 	'lookbehind assertion is not fixed length'
#> 	at '(?<!:/?)/)'
#> Error in sub("^.*?(?=(?<!:/?)/)", "", redirect_url, perl = TRUE): invalid regular expression '^.*?(?=(?<!:/?)/)'
fp$add_auth(google, "google_auth")
#> Error: object 'google' not found

# Add authentication to different paths
fp$add_auth_handler("get", "/require_basic", basic_auth)

fp$add_auth_handler("get", "/require_basic_and_key", basic_auth && key_auth)

fp$add_auth_handler(
  "get",
  "/require_google_or_the_others",
  google_auth || (basic_auth && key_auth)
)

# Add plugin to fiery app
app <- fiery::Fire$new()

# First add the firesale plugin as it is required
fs <- firesale::FireSale$new(storr::driver_environment(new.env()))
app$attach(fs)

# Then add the fireproof plugin
app$attach(fp)