Authentication based on shared secret

This authentication scheme is based on a mutually shared secret between the server and the client. The client provides this secret either as a header or in a cookie, and the server verifies the authenticity of the secret. Like with basic authentication, this scheme relies on additional technology like HTTPS/SSL to make it secure since the secret can otherwise easily be extracted from the request by man-in-the-middle attack.

auth_key(key, secret, user_info = NULL, cookie = TRUE, name = "KeyAuth")

Arguments

key

The name of the header or cookie to store the secret under

secret

The secret to check for. Either a single string with the secret or a function that will be called with the arguments key, request, and response and returns TRUE if its a valid secret (useful if you have multiple or rotating secrets). If the function returns a character vector it is considered to be authenticated and the return value will be understood as scopes the user is granted. Make sure never to store secrets in plain text and avoid checking them into version control.

user_info

A function to extract user information from the username. It is called with two arguments: key and setter, the first being the key used for the successful authentication, the second being a function that must be called in the end with the relevant information (see user_info()).

cookie

Boolean. Should the secret be transmitted as a cookie. If FALSE it is expected to be transmitted as a header.

name

The name of the authentication

Value

A AuthKey object

Details

This authentication is not a classic HTTP authentication scheme and thus doesn't return a 401 response with a WWW-Authenticate header. Instead it returns a 400 response unless another authenticator has already set the response status to something else.

User information

auth_key() automatically adds user information after authentication. By default it will set the provider field to "local". Further, it will set the scopes field to any scopes returned by the authenticator function.

Since Key based authentication is seldom used with user specific keys it is unlikely that it makes sense to populate the information any further.

Examples

# Create an authenticator of dubious quality
key <- auth_key(
  key = "my-key-location",
  secret = "SHHH!!DONT_TELL_ANYONE"
)

# Add it to a fireproof plugin
fp <- Fireproof$new()
fp$add_auth(key, "key_auth")

# Use it in an endpoint
fp$add_auth_handler("get", "/*", key_auth)